+++ src/onegov.election_day/onegov/election_day/formats/election/wabsti_majorz.py
@@ -5,7 +5,6 @@ from onegov.election_day import _
 from onegov.election_day.formats.common import EXPATS
 from onegov.election_day.formats.common import FileImportError
 from onegov.election_day.formats.common import load_csv
-from onegov.election_day.utils import clear_election
 from uuid import uuid4
 
 
@@ -225,7 +224,7 @@ def import_election_wabsti_majorz(
         return errors
 
     if results:
-        clear_election(election)
+        election.clear_results()
 
         election.number_of_mandates = mandates
         election.counted_entities = len(results)