+++ src/onegov.election_day/onegov/election_day/formats/election/wabstic_majorz.py
@@ -5,7 +5,6 @@ from onegov.election_day import _
 from onegov.election_day.formats.common import EXPATS
 from onegov.election_day.formats.common import FileImportError
 from onegov.election_day.formats.common import load_csv
-from onegov.election_day.utils import clear_election
 from uuid import uuid4
 
 HEADERS_WM_WAHL = (
@@ -338,7 +337,7 @@ def import_election_wabstic_majorz(
         return errors
 
     if added_results:
-        clear_election(election)
+        election.clear_results()
 
         election.absolute_majority = absolute_majority
         election.counted_entities = sum([