+++ src/onegov.election_day/onegov/election_day/formats/vote/wabstim.py
@@ -3,7 +3,6 @@ from onegov.ballot import BallotResult
 from onegov.election_day import _
 from onegov.election_day.formats.common import FileImportError
 from onegov.election_day.formats.common import load_csv
-from onegov.election_day.utils import clear_vote
 from onegov.election_day.utils import guessed_group
 
 
@@ -159,7 +158,7 @@ def import_vote_wabstim(vote, entities, complex, file, mimetype):
     ):
         return [FileImportError(_("No data found"))]
 
-    clear_vote(vote)
+    vote.clear_results()
 
     for ballot_type in used_ballot_types:
         remaining = (